requirements.txt
# Core Backend
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
itsdangerous==2.1.2
python-dotenv==1.0.0

# Database
pymongo==4.6.0

# LangChain & LangGraph (Compatible versions)
langchain==0.3.7
langchain-google-genai==2.0.4
langgraph==0.2.34
langchain_huggingface==0.1.0

# Redis
upstash-redis==0.15.0

# Vector Database
pinecone-client==5.0.1

# File Processing
unstructured[docx,pdf,pptx,xlsx]>=0.14.10
pandas==2.1.4
psutil==5.9.8

# Utilities
pydantic>=2.7.4
pydantic-settings>=2.0.0
pydantic[email]>=2.7.4

# Testing
requests==2.31.0

# Security
bcrypt==4.1.2

